cbp_cam/requirements.txt
opencv_python
torchvision
numpy
matplotlib
Flask
scikit_image
Pillow
